User story #1481: Update: Opsplit JS-logik'en for DataIO og FBS.
Fixed build errors after split of JS.

git-svn-id: https://svn.dbc.dk/repos/ocb-tools/trunk@83598 b87a6194-a221-0410-ad6f-3b8ec35bba13

+++ libraries/ocb-test-engine/src/main/java/dk/dbc/ocbtools/testengine/executors/Solr.java
@@ -43,9 +43,11 @@ public class Solr {
 
         String url = "";
         try {
-            url = settings.getProperty( "solr.url" );
-            SolrServer solr = new CommonsHttpSolrServer( url );
-            solr.deleteByQuery("*:*");
+            if( settings.containsKey( "solr.url" ) ) {
+                url = settings.getProperty( "solr.url" );
+                SolrServer solr = new CommonsHttpSolrServer( url );
+                solr.deleteByQuery( "*:*" );
+            }
         }
         catch( IOException | SolrServerException ex ) {
             output.error( "Unable to clear the solr index at {}: {}", url, ex.getMessage() );
